started compartmentalizing tensorflow graph into glm general and noise model specific subgraphs
model is complete, currently working on estimator graph. estimator, hessian and jacobian are yet to come.
