+++ batchglm/train/tf/base_glm/__init__.py
@@ -1,0 +1,2 @@
+from .estimator_graph import GradientGraph, NewtonGraph, TrainerGraph
+from .model import ESTIMATOR_PARAMS, _ProcessModel, _ModelVars
