+++ batchglm/train/tf/base_glm/external.py
@@ -1,0 +1,4 @@
+import batchglm.train.tf.ops as op_utils
+import batchglm.train.tf.train as train_utils
+
+from batchglm.models.nb_glm import AbstractEstimator
