+++ batchglm/train/tf/nb_glm/estimator_graph.py
@@ -10,9 +10,9 @@ try:
 except ImportError:
     anndata = None
 
-from .model import BasicModelGraph, ModelVars
-from .model import param_bounds
+from .model import BasicModelGraph, ModelVars, ProcessModel
 
+from .external import GradientGraph, NewtonGraph, TrainerGraph
 from .external import TFEstimatorGraph
 from .external import train_utils, op_utils
 from .external import pkg_constants
@@ -218,7 +218,8 @@ class FullDataModelGraph:
         self.neg_jac_train = jacobian_train.neg_jac
         self.neg_hessian_train = hessians_train.neg_hessian
 
-class EstimatorGraph(TFEstimatorGraph):
+
+class EstimatorGraph(GradientGraph, NewtonGraph, TrainerGraph, TFEstimatorGraph):
     """
     The estimator graphs are all graph necessary to perform parameter updates and to
     summarise a current parameter estimate.
@@ -305,7 +306,8 @@ class EstimatorGraph(TFEstimatorGraph):
         :param extended_summary:
         :param dtype: Precision used in tensorflow.
         """
-        super().__init__(graph)
+        TFEstimatorGraph.__init__(graph)
+
         self.num_observations = num_observations
         self.num_features = num_features
         self.num_design_loc_params = num_design_loc_params
@@ -436,326 +438,47 @@ class EstimatorGraph(TFEstimatorGraph):
                 r = full_data_model.r
                 sigma2 = full_data_model.sigma2
 
-            # ### management
-            with tf.name_scope("training"):
-                logger.debug(" ** Build training graphs")
-                global_step = tf.train.get_or_create_global_step()
+
                 idx_nonconverged = np.where(model_vars.converged == False)[0]
 
-                # Define newton-rhapson updates and gradients depending on termination type
-                # and depending on whether data is batched.
-                # The formed have to be distinguished because gradients and parameter updates
-                # are set to zero (or not computed in the case of newton-raphson) for
-                # converged features if feature-wise termination is chosen.
-                # The latter have to be distinguished as there are different jacobians
-                # and hessians for the full and the batched data.
-                if termination_type == "by_feature":
-                    if provide_optimizers["nr"]:
-                        logger.debug(" ** Build newton training graph: by_feature, train mu and r")
-                        # Full data model by gene:
-                        param_grad_vec = full_data_model.neg_jac_train
-                        # Compute parameter update for non-converged gene only.
-                        delta_t_bygene_nonconverged = tf.squeeze(tf.matrix_solve_ls(
-                            tf.gather(
-                                full_data_model.neg_hessian_train,
-                                indices=idx_nonconverged,
-                                axis=0),
-                            tf.expand_dims(
-                                tf.gather(
-                                    param_grad_vec,
-                                    indices=idx_nonconverged,
-                                    axis=0)
-                                , axis=-1),
-                            fast=False
-                        ), axis=-1)
-                        # Write parameter updates into matrix of size of all parameters which
-                        # contains zero entries for updates of already converged genes.
-                        delta_t_bygene_full = tf.concat([
-                            tf.gather(delta_t_bygene_nonconverged,
-                                      indices=np.where(idx_nonconverged == i)[0],
-                                      axis=0)
-                            if not model_vars.converged[i]
-                            else tf.zeros([1, param_grad_vec.shape[1]])
-                            for i in range(model_vars.n_features)
-                        ], axis=0)
-                        nr_update_full = tf.transpose(delta_t_bygene_full)
+                self.model_vars = model_vars
+                self.batch_model = batch_model
+                self.learning_rate = learning_rate
+                self.loss = batch_loss
 
-                        # Batched data model by gene:
-                        param_grad_vec_batched = batch_jac.neg_jac
-                        # Compute parameter update for non-converged gene only.
-                        delta_batched_t_bygene_nonconverged = tf.squeeze(tf.matrix_solve_ls(
-                            tf.gather(
-                                batch_hessians.neg_hessian,
-                                indices=idx_nonconverged,
-                                axis=0),
-                            tf.expand_dims(
-                                tf.gather(
-                                    param_grad_vec_batched,
-                                    indices=idx_nonconverged,
-                                    axis=0)
-                                , axis=-1),
-                            fast=False
-                        ), axis=-1)
-                        # Write parameter updates into matrix of size of all parameters which
-                        # contains zero entries for updates of already converged genes.
-                        delta_batched_t_bygene_full = tf.concat([
-                            tf.gather(delta_batched_t_bygene_nonconverged,
-                                      indices=np.where(idx_nonconverged == i)[0],
-                                      axis=0)
-                            if not model_vars.converged[i]
-                            else tf.zeros([1, param_grad_vec.shape[1]])
-                            for i in range(model_vars.n_features)
-                        ], axis=0)
-                        nr_update_batched = tf.transpose(delta_batched_t_bygene_full)
-                    else:
-                        nr_update_full = None
-                        nr_update_batched = None
+                self.mu = mu
+                self.r = r
+                self.sigma2 = sigma2
 
-                    gradients_batch_all = tf.transpose(batch_jac.neg_jac)
-                    gradients_batch = tf.concat([
-                        # tf.gradients(batch_model.norm_neg_log_likelihood,
-                        #             model_vars.params_by_gene[i])[0]
-                        tf.expand_dims(gradients_batch_all[:, i], axis=-1)
-                        if not model_vars.converged[i]
-                        else tf.zeros([param_grad_vec.shape[1], 1], dtype=dtype)
-                        for i in range(model_vars.n_features)
-                    ], axis=1)
+                self.batch_probs = batch_model.probs
+                self.batch_log_probs = batch_model.log_probs
+                self.batch_log_likelihood = batch_model.norm_log_likelihood
 
-                    gradients_full_all = tf.transpose(full_data_model.neg_jac_train)
-                    gradients_full = tf.concat([
-                        # tf.gradients(full_data_model.norm_neg_log_likelihood,
-                        #             model_vars.params_by_gene[i])[0]
-                        tf.expand_dims(gradients_full_all[:, i], axis=-1)
-                        if not model_vars.converged[i]
-                        else tf.zeros([param_grad_vec.shape[1], 1], dtype=dtype)
-                        for i in range(model_vars.n_features)
-                    ], axis=1)
-                elif termination_type == "global":
-                    if provide_optimizers["nr"]:
-                        # Full data model:
-                        param_grad_vec = full_data_model.neg_jac_train
-                        delta_t = tf.squeeze(tf.matrix_solve_ls(
-                            full_data_model.neg_hessian_train,
-                            # (full_data_model.hessians + tf.transpose(full_data_model.hessians, perm=[0, 2, 1])) / 2,  # symmetrization, don't need this with closed forms
-                            tf.expand_dims(param_grad_vec, axis=-1),
-                            fast=False
-                        ), axis=-1)
-                        nr_update_full = tf.transpose(delta_t)
+                self.sample_selection = sample_selection
+                self.full_data_model = full_data_model
 
-                        # Batched data model:
-                        param_grad_vec_batched = batch_jac.neg_jac
-                        delta_batched_t = tf.squeeze(tf.matrix_solve_ls(
-                            batch_hessians.neg_hessian,
-                            tf.expand_dims(param_grad_vec_batched, axis=-1),
-                            fast=False
-                        ), axis=-1)
-                        nr_update_batched = tf.transpose(delta_batched_t)
-                    else:
-                        nr_update_full = None
-                        nr_update_batched = None
+                self.full_loss = full_data_loss
+                self.full_gradient = full_gradient
 
-                    gradients_batch = tf.transpose(batch_jac.neg_jac)
-                    gradients_full = tf.transpose(full_data_model.neg_jac_train)
-                else:
-                    raise ValueError("convergence_type %s not recognized." % termination_type)
+                self.hessians = full_data_model.hessian
+                self.fisher_inv = fisher_inv
 
-                # Pad gradient and update vectors to receive update tensors that match
-                # the shape of model_vars.params.
-                if train_mu:
-                    if train_r:
-                        gradients_batch_all = gradients_batch
-                        gradients_full_all = gradients_full
-                        if provide_optimizers["nr"]:
-                            nr_update_batched_all = nr_update_batched
-                            nr_update_full_all = nr_update_full
-                        else:
-                            nr_update_batched_all = None
-                            nr_update_full_all = None
-                    else:
-                        gradients_batch_all = tf.concat([
-                            gradients_batch,
-                            tf.zeros_like(model_vars.b)
-                        ], axis=0)
-                        gradients_full_all = tf.concat([
-                            gradients_full,
-                            tf.zeros_like(model_vars.b)
-                        ], axis=0)
-                        if provide_optimizers["nr"]:
-                            nr_update_batched_all = tf.concat([
-                                nr_update_batched,
-                                tf.zeros_like(model_vars.b)
-                            ], axis=0)
-                            nr_update_full_all = tf.concat([
-                                nr_update_full,
-                                tf.zeros_like(model_vars.b)
-                            ], axis=0)
-                        else:
-                            nr_update_batched_all = None
-                            nr_update_full_all = None
-                elif train_r:
-                    gradients_batch_all = tf.concat([
-                        tf.zeros_like(model_vars.a),
-                        gradients_batch
-                    ], axis=0)
-                    gradients_full_all = tf.concat([
-                        tf.zeros_like(model_vars.a),
-                        gradients_full
-                    ], axis=0)
-                    if provide_optimizers["nr"]:
-                        nr_update_batched_all = tf.concat([
-                            tf.zeros_like(model_vars.a),
-                            nr_update_batched
-                        ], axis=0)
-                        nr_update_full_all = tf.concat([
-                            tf.zeros_like(model_vars.a),
-                            nr_update_full
-                        ], axis=0)
-                    else:
-                        nr_update_batched_all = None
-                        nr_update_full_all = None
-                else:
-                    logger.info("No training necessary; returning")
-                    return
+                self.idx_nonconverged = idx_nonconverged
 
-                # Create trainers that produce training operations.
-                trainer_batch = train_utils.MultiTrainer(
-                    variables=[model_vars.params],
-                    gradients=[(gradients_batch_all, model_vars.params)],
-                    newton_delta=nr_update_batched_all,
-                    learning_rate=learning_rate,
-                    global_step=global_step,
-                    apply_gradients=lambda grad: tf.where(tf.is_nan(grad), tf.zeros_like(grad), grad),
+                GradientGraph.__init__(
+                    termination_type=termination_type,
+                    train_mu=train_mu,
+                    train_r=train_r
+                )
+                NewtonGraph.__init__(
+                    termination_type=termination_type,
                     provide_optimizers=provide_optimizers,
-                    name="batch_data_trainers"
+                    train_mu=train_mu,
+                    train_r=train_r
                 )
-
-                trainer_full = train_utils.MultiTrainer(
-                    variables=[model_vars.params],
-                    gradients=[(gradients_full_all, model_vars.params)],
-                    newton_delta=nr_update_full_all,
-                    learning_rate=learning_rate,
-                    global_step=global_step,
-                    apply_gradients=lambda grad: tf.where(tf.is_nan(grad), tf.zeros_like(grad), grad),
-                    provide_optimizers=provide_optimizers,
-                    name="full_data_trainers"
+                TrainerGraph.__init__(
+                    feature_isnonzero=feature_isnonzero
                 )
-
-                # Set up model gradient computation:
-                with tf.name_scope("batch_gradient"):
-                    logger.debug(" ** Build training graph: batched data model gradients")
-                    batch_gradient = trainer_batch.plain_gradient_by_variable(model_vars.params)
-                    batch_gradient = tf.reduce_sum(tf.abs(batch_gradient), axis=0)
-
-                    # batch_gradient = tf.add_n(
-                    #     [tf.reduce_sum(tf.abs(grad), axis=0) for (grad, var) in batch_trainers.gradient])
-
-                with tf.name_scope("full_gradient"):
-                    logger.debug(" ** Build training graph: full data model gradients")
-                    # use same gradient as the optimizers
-                    full_gradient = trainer_full.plain_gradient_by_variable(model_vars.params)
-                    full_gradient = tf.reduce_sum(tf.abs(full_gradient), axis=0)
-
-                    # # the analytic Jacobian
-                    # full_gradient = tf.reduce_sum(full_data_model.neg_jac, axis=0)
-                    # full_gradient = tf.add_n(
-                    #     [tf.reduce_sum(tf.abs(grad), axis=0) for (grad, var) in full_data_trainers.gradient])
-
-                # # ### BFGS implementation using SciPy L-BFGS
-                # with tf.name_scope("bfgs"):
-                #     feature_idx = tf.placeholder(dtype="int64", shape=())
-                #
-                #     X_s = tf.gather(X, feature_idx, axis=1)
-                #     a_s = tf.gather(a, feature_idx, axis=1)
-                #     b_s = tf.gather(b, feature_idx, axis=1)
-                #
-                #     model = BasicModelGraph(X_s, design_loc, design_scale, a_s, b_s, size_factors=size_factors)
-                #
-                #     trainer = tf.contrib.opt.ScipyOptimizerInterface(
-                #         model.loss,
-                #         method='L-BFGS-B',
-                #         options={'maxiter': maxiter})
-
-            with tf.name_scope("init_op"):
-                logger.debug(" ** Build training graph: initialization operation")
-                init_op = tf.global_variables_initializer()
-
-            # ### output values:
-            #       override all-zero features with lower bound coefficients
-            with tf.name_scope("output"):
-                logger.debug(" ** Build training graph: output")
-                bounds_min, bounds_max = param_bounds(dtype)
-
-                param_nonzero_a = tf.broadcast_to(feature_isnonzero, [num_design_loc_params, num_features])
-                alt_a = tf.concat([
-                    # intercept
-                    tf.broadcast_to(bounds_min["a"], [1, num_features]),
-                    # slope
-                    tf.zeros(shape=[num_design_loc_params - 1, num_features], dtype=model_vars.a.dtype),
-                ], axis=0, name="alt_a")
-                a = tf.where(
-                    param_nonzero_a,
-                    model_vars.a,
-                    alt_a,
-                    name="a"
-                )
-
-                param_nonzero_b = tf.broadcast_to(feature_isnonzero, [num_design_scale_params, num_features])
-                alt_b = tf.concat([
-                    # intercept
-                    tf.broadcast_to(bounds_max["b"], [1, num_features]),
-                    # slope
-                    tf.zeros(shape=[num_design_scale_params - 1, num_features], dtype=model_vars.b.dtype),
-                ], axis=0, name="alt_b")
-                b = tf.where(
-                    param_nonzero_b,
-                    model_vars.b,
-                    alt_b,
-                    name="b"
-                )
-
-        logger.debug(" ** Build training graph: setting class attributes")
-        self.fetch_fn = fetch_fn
-        self.model_vars = model_vars
-        self.batch_model = batch_model
-
-        self.learning_rate = learning_rate
-        self.loss = batch_loss
-
-        self.trainer_batch = trainer_batch
-        self.trainer_full = trainer_full
-        self.global_step = global_step
-
-        self.gradient = batch_gradient
-        # self.gradient_a = batch_gradient_a
-        # self.gradient_b = batch_gradient_b
-
-        self.train_op = trainer_full.train_op_GD
-
-        self.init_ops = []
-        self.init_op = init_op
-
-        # # ### set up class attributes
-        self.a = a
-        self.b = b
-        assert (self.a.shape == (num_design_loc_params, num_features))
-        assert (self.b.shape == (num_design_scale_params, num_features))
-
-        self.mu = mu
-        self.r = r
-        self.sigma2 = sigma2
-
-        self.batch_probs = batch_model.probs
-        self.batch_log_probs = batch_model.log_probs
-        self.batch_log_likelihood = batch_model.norm_log_likelihood
-
-        self.sample_selection = sample_selection
-        self.full_data_model = full_data_model
-
-        self.full_loss = full_data_loss
-        self.full_gradient = full_gradient
-
-        self.hessians = full_data_model.hessian
-        self.fisher_inv = fisher_inv
 
         with tf.name_scope('summaries'):
             tf.summary.histogram('a', model_vars.a)
