+++ batchglm/train/tf/nb_glm/external.py
@@ -6,6 +6,9 @@ from batchglm.models.nb_glm.utils import closedform_nb_glm_logmu, closedform_nb_
 import batchglm.train.tf.ops as op_utils
 import batchglm.train.tf.train as train_utils
 from batchglm.train.tf.base import TFEstimatorGraph, MonitoredTFEstimator
+from batchglm.train.tf.base_glm import GradientGraph, NewtonGraph, TrainerGraph
+
+from batchglm.train.tf.base_glm import ESTIMATOR_PARAMS, _ProcessModel, _ModelVars
 
 import batchglm.utils.random as rand_utils
 from batchglm.utils.linalg import groupwise_solve_lm
